Move NewTenant module to new location
