+++ Resources/Private/Templates/Backend/NewTenant/Error.html
@@ -14,6 +14,6 @@
 
 <h1><f:translate key="newTenant.module" /></h1>
 
-<f:render partial="InfoBox" arguments="{title: '{f:translate(key: \'newTenant.wrongPageType\')}', message: '{f:translate(key: \'newTenant.wrongPageTypeMsg\')}', alert: 'danger'}" />
+<f:render partial="Backend/InfoBox" arguments="{title: '{f:translate(key: \'newTenant.wrongPageType\')}', message: '{f:translate(key: \'newTenant.wrongPageTypeMsg\')}', alert: 'danger'}" />
 
 </html>
