+++ Resources/Private/Templates/Backend/NewTenant/Index.html
@@ -20,11 +20,11 @@
     <f:if condition="{recordInfos.formats.numCurrent} >= {recordInfos.formats.numDefault}">
         <f:then>
             <f:variable name="formatsAvailable" value="1" />
-            <f:render partial="InfoBox" arguments="{title: '{f:translate(key: \'newTenant.formatOkay\')}', message: '{f:translate(key: \'newTenant.formatOkayMsg\')}', alert: 'success'}" />
+            <f:render partial="Backend/InfoBox" arguments="{title: '{f:translate(key: \'newTenant.formatOkay\')}', message: '{f:translate(key: \'newTenant.formatOkayMsg\')}', alert: 'success'}" />
         </f:then>
         <f:else>
             <f:variable name="formatsAvailable" value="0" />
-            <f:render partial="InfoBox" arguments="{title: '{f:translate(key: \'newTenant.formatNotOkay\')}', message: '{f:translate(key: \'newTenant.formatNotOkayMsg\')}', alert: 'danger'}" />
+            <f:render partial="Backend/InfoBox" arguments="{title: '{f:translate(key: \'newTenant.formatNotOkay\')}', message: '{f:translate(key: \'newTenant.formatNotOkayMsg\')}', alert: 'danger'}" />
             <f:form action="addFormat" name="format">
                 <f:form.submit name="submit" value="{f:translate(key: 'newTenant.createFormat')}" class="btn btn-primary"/>
             </f:form>
@@ -37,10 +37,10 @@
 
     <f:if condition="{recordInfos.structures.numCurrent} >= {recordInfos.structures.numDefault}">
         <f:then>
-            <f:render partial="InfoBox" arguments="{title: '{f:translate(key: \'newTenant.structureOkay\')}', message: '{f:translate(key: \'newTenant.structureOkayMsg\')}', alert: 'success'}" />
+            <f:render partial="Backend/InfoBox" arguments="{title: '{f:translate(key: \'newTenant.structureOkay\')}', message: '{f:translate(key: \'newTenant.structureOkayMsg\')}', alert: 'success'}" />
         </f:then>
         <f:else>
-            <f:render partial="InfoBox" arguments="{title: '{f:translate(key: \'newTenant.structureNotOkay\')}', message: '{f:translate(key: \'newTenant.structureNotOkayMsg\')}', alert: 'danger'}" />
+            <f:render partial="Backend/InfoBox" arguments="{title: '{f:translate(key: \'newTenant.structureNotOkay\')}', message: '{f:translate(key: \'newTenant.structureNotOkayMsg\')}', alert: 'danger'}" />
             <f:form action="addStructure" name="structure">
                 <f:form.submit name="submit" value="{f:translate(key: 'newTenant.createStructure')}" class="btn btn-primary"/>
             </f:form>
@@ -52,10 +52,10 @@
 <div style="padding: 5px;">
     <f:if condition="{recordInfos.metadata.numCurrent} >= {recordInfos.metadata.numDefault}">
         <f:then>
-            <f:render partial="InfoBox" arguments="{title: '{f:translate(key: \'newTenant.metadataOkay\')}', message: '{f:translate(key: \'newTenant.metadataOkayMsg\')}', alert: 'success'}" />
+            <f:render partial="Backend/InfoBox" arguments="{title: '{f:translate(key: \'newTenant.metadataOkay\')}', message: '{f:translate(key: \'newTenant.metadataOkayMsg\')}', alert: 'success'}" />
         </f:then>
         <f:else>
-            <f:render partial="InfoBox" arguments="{title: '{f:translate(key: \'newTenant.metadataNotOkay\')}', message: '{f:translate(key: \'newTenant.metadataNotOkayMsg\')}', alert: 'danger'}" />
+            <f:render partial="Backend/InfoBox" arguments="{title: '{f:translate(key: \'newTenant.metadataNotOkay\')}', message: '{f:translate(key: \'newTenant.metadataNotOkayMsg\')}', alert: 'danger'}" />
             <f:form action="addMetadata" name="metadata">
                 <f:if condition="{formatsAvailable} == '0'">
                     <f:then>
@@ -73,10 +73,10 @@
 <div style="padding: 5px;">
     <f:if condition="{recordInfos.solrcore.numCurrent} >= 1">
         <f:then>
-            <f:render partial="InfoBox" arguments="{title: '{f:translate(key: \'newTenant.solrcoreOkay\')}', message: '{f:translate(key: \'newTenant.solrcoreOkayMsg\')}', alert: 'success'}" />
+            <f:render partial="Backend/InfoBox" arguments="{title: '{f:translate(key: \'newTenant.solrcoreOkay\')}', message: '{f:translate(key: \'newTenant.solrcoreOkayMsg\')}', alert: 'success'}" />
         </f:then>
         <f:else>
-            <f:render partial="InfoBox" arguments="{title: '{f:translate(key: \'newTenant.solrcoreNotOkay\')}', message: '{f:translate(key: \'newTenant.solrcoreNotOkayMsg\')}', alert: 'danger'}" />
+            <f:render partial="Backend/InfoBox" arguments="{title: '{f:translate(key: \'newTenant.solrcoreNotOkay\')}', message: '{f:translate(key: \'newTenant.solrcoreNotOkayMsg\')}', alert: 'danger'}" />
             <f:form action="addSolrCore" name="solr">
                 <f:form.submit name="submit" value="{f:translate(key: 'newTenant.createSolrCore')}" class="btn btn-primary" />
             </f:form>
