+++ ext_localconf.php
@@ -300,36 +300,6 @@ $GLOBALS['TYPO3_CONF_VARS']['SC_OPTIONS']['ext/install']['update'][\Kitodo\Dlf\U
     ]
 );
 
-\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addTypoScriptSetup(
-    'module.tx_dlf {
-        view {
-            templateRootPaths {
-                0 = EXT:dlf/Resources/Private/Backend/Templates/
-                1 = {$module.tx_dlf.view.templateRootPath}
-            }
-            partialRootPaths {
-                0 = EXT:dlf/Resources/Private/Backend/Partials/
-                1 = {$module.tx_dlf.view.partialRootPath}
-            }
-            layoutRootPaths {
-                0 = EXT:dlf/Resources/Private/Backend/Layouts/
-                1 = {$module.tx_dlf.view.layoutRootPath}
-            }
-        }
-    }'
-);
-\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addTypoScriptConstants(
-    'module.tx_dlf {
-        view {
-            # cat=module.tx_dlf/file; type=string; label=Path to template root (BE)
-            templateRootPath = EXT:dlf/Resources/Private/Backend/Templates/
-            # cat=module.tx_dlf/file; type=string; label=Path to template partials (BE)
-            partialRootPath = EXT:dlf/Resources/Private/Backend/Partials/
-            # cat=module.tx_dlf/file; type=string; label=Path to template layouts (BE)
-            layoutRootPath = EXT:dlf/Resources/Private/Backend/Layouts/
-        }
-    }'
-);
 
 // Register a node in ext_localconf.php
 $GLOBALS['TYPO3_CONF_VARS']['SYS']['formEngine']['nodeRegistry'][1638809996] = [
